requirements.txt
streamlit
pandas
jupyterlab
python-dotenv
plotly
fpdf